Separate code by area of concern for manageability
