--- templates/style.css
@@ -1,4 +1,0 @@
-img {
-    width: 320px;
-    height: 280px;
-}
